fixed blue initialisation

+++ improvements.py
@@ -31,7 +31,7 @@ class BlurLayer(nn.Module):
         self.kernel_size = kernel_size
         
         # Train for the sqrt of the sigma to keep sigma positie when calculating Gaussian
-        self.sqrt_sigma = nn.Parameter( torch.tensor(sqrt_sigma_initial) )
+        self.sqrt_sigma = nn.Parameter( torch.tensor(sqrt_sigma_initial, dtype=torch.float32) )
 
     def forward(self, x):
         # Applying blurring to x using a learned level of blurring
@@ -109,7 +109,7 @@ class ImprovedMrCNN(nn.Module):
 
         self.stream1 = ImprovedMrCNNStream(blur_kernel_size = 3, blur_sqrt_sigma_initial = 0.5)
         self.stream2 = ImprovedMrCNNStream(blur_kernel_size = 5, blur_sqrt_sigma_initial = 0.75)
-        self.stream3 = ImprovedMrCNNStream(blur_kernel_size = 7, blur_sqrt_sigma_initial = 1)
+        self.stream3 = ImprovedMrCNNStream(blur_kernel_size = 7, blur_sqrt_sigma_initial = 1.0)
 
         self.fc = nn.Linear(in_features=(512 * 3), out_features=512)
